Correction to not allow repeated chromosome in the solution of the multi-objective algorithm and implementation of the analysis of the objectives of the individuals of the solution of the multi-objective algorithm

+++ src/br/ufu/scheduling/model/Metrics.java
@@ -267,7 +267,7 @@ public class Metrics implements Cloneable {
 	}
 
 	private Double calculateFitnessForMetric(Configuration config, MetricType metricType) {
-	    return (double) config.getTransformedObjectiveValue(getMetricValue(metricType));
+	    return config.getTransformedObjectiveValue(getMetricValue(metricType));
 	}
 
 	private void calculateFitness(Configuration config) {
